login - logout -> ok

+++ partials/header.php
@@ -5,7 +5,7 @@
   require_once(__DIR__.'/../config/config.php');
   // Inclusion du fichier database
   require_once(__DIR__.'/../config/database.php');
-  
+  session_start();
 ?>
 
 <!DOCTYPE html>
@@ -15,7 +15,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="description" content="">
     <meta name="author" content="">
-    <link rel="icon" href="assets/img/favicon.ico">
+    
 
     <title>
       <?php
@@ -49,19 +49,36 @@
       </button>
 
       <div class="collapse navbar-collapse" id="navbar-webflix">
-        <ul class="navbar-nav mr-auto">
-          <li class="nav-item <?php echo ($currentPageUrl === 'index') ? 'active' : ''; ?>">
-            <a class="nav-link" href="index.php">Accueil</a>
-          </li>          
-          <li class="nav-item <?= ($currentPageUrl === 'movie_add') ? 'active' : ''; ?>">
-            <a class="nav-link" href="movie_add.php">Ajouter un film</a>
-          </li>
-          <li class="nav-item <?= ($currentPageUrl === 'movie_modify_delete') ? 'active' : ''; ?>">
-            <a class="nav-link" href="movie_modify_delete.php">Modifier/Supprimer</a>
-          </li>
+        <ul class="navbar-nav mr-auto">        
+            <li class="nav-item <?php echo ($currentPageUrl === 'index') ? 'active' : ''; ?>">
+              <a class="nav-link" href="index.php">Accueil</a>
+            </li>           
+          <?php if(!empty($_SESSION)) {?>
+            <li class="nav-item <?= ($currentPageUrl === 'movie_add') ? 'active' : ''; ?>">
+              <a class="nav-link" href="movie_add.php">Ajouter un film</a>
+            </li>
+            <li class="nav-item <?= ($currentPageUrl === 'movie_modify_delete') ? 'active' : ''; ?>">
+              <a class="nav-link" href="movie_modify_delete.php">Modifier/Supprimer</a>
+            </li>
+          <?php } ?>
         </ul>
-        <div class="justify-content-end">
-          <button type="button" class="btn btn-outline-success" data-toggle="modal" data-target="#modalConnexion">Connexion</button>
+        <div class="d-flex justify-content-end">
+        <?php 
+        if(empty($_SESSION)) { ?>
+          <button type="button" class="btn btn-outline-success" data-toggle="modal" data-target="#modalConnexion">Connexion</button>          
+        <?php }else{ ?>
+          <div class="dropdown">
+            <a class="btn btn-secondary dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+              <?= $_SESSION['user'] ?>
+            </a>
+            <div class="dropdown-menu" aria-labelledby="dropdownMenuLink">
+              <a class="dropdown-item" href="#">Action</a>
+              <a class="dropdown-item" href="#">Another action</a>
+              <a class="btn btn-danger deco" href="logout.php">Déconnexion</a>
+            </div>
+          </div>
+        <?php }
+        ?>
         </div>
           
       </div>
@@ -85,36 +102,37 @@
         </button>
       </div>
       <div class="modal-body">
-        <form id="signin" class="form-signin" action="user_connect.php">                    
-          <label for="inputEmail" class="sr-only">Email address</label>
-          <input type="email" id="inputEmail" class="form-control" placeholder="Email address" required autofocus>
+        <form id="signin" class="form-signin" action="login.php" method="post">                    
+          <label for="inputEmail" class="sr-only" >Email address</label>
+          <input type="email" name="mailAdress" id="inputEmail" class="form-control" placeholder="Email address" required autofocus>
           <label for="inputPassword" class="sr-only">Password</label>
-          <input type="password" id="inputPassword" class="form-control" placeholder="Password" required>
+          <input type="password" name="pwdConnect" id="inputPassword" class="form-control" placeholder="Password" required>
           <div class="checkbox mb-3">
             <label>
               <input type="checkbox" value="remember-me"> Remember me
             </label>
           </div>
-          <button class="btn btn-lg btn-success btn-block" type="submit">Connexion</button>
+          <input type="submit" class="btn btn-lg btn-success btn-block" value="Connexion">
           <p class="mt-5 mb-3 text-muted">&copy; 2018</p>
         </form>
-        <form id="create" class="form-signin create" action="user_add.php">
+        <form id="create" class="form-signin create" action="register.php" method="post">
 
           <label for="mail" class="sr-only">Email address</label>
-          <input type="email" id="mail" class="form-control" placeholder="Email address" required autofocus>
+          <input type="email" id="mail" name="inputMail" class="form-control" placeholder="Email address" required autofocus>
 
           <label for="username" class="sr-only">Username</label>
-          <input type="text" id="username" class="form-control" placeholder="Username" required>
+          <input type="text" id="username" name="inputUser" class="form-control" placeholder="Username" required>
 
           <label for="pwd" class="sr-only">Password</label>
-          <input type="password" id="pwd" class="form-control" placeholder="Password" required>
+          <input type="password" id="pwd" name="inputPwd" class="form-control" placeholder="Password" required>
 
           <label for="Confirm_pwd" class="sr-only">Confirm password</label>
-          <input type="password" id="Confirm_pwd" class="form-control" placeholder="Confirm Password" required>
+          <input type="password" id="Confirm_pwd" name="inputCoPwd" class="form-control" placeholder="Confirm Password" required>
           
           <input class="btn btn-lg btn-primary btn-block" type="submit" value="Créer un compte">
           <p class="mt-5 mb-3 text-muted">&copy; 2018</p>
         </form>
+        
       </div>
       
     </div>
